gfx2d: pre-multiply the source with the source color in blend commands
When the source color is not (1, 1, 1, 1), create a intermediate surface
with ARGB8888 pixel format once for all, then multiply the original
source surface with the source color and store the result into the
intermediate surface.

Finally, blend the intermediate surface, in replacement of the original
source surface, with the destination surface and store the result into
the target surface as usual.

N.B: don't try to use the GFX2D pre-mult feature with BLEND_WD5.SPRE bit
and BLEND_WD0.REG constant as it doesn't work as you may expect.
For instance, if the BLEND_WD5.SCFACT or BLEND_WD5.SAFACT are ONE or
ZERO,
no pre-mult is done at all.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Reviewed-by: Ludovic Desroches <ludovic.desroches@microchip.com>